refactor : mock 데이터 fetching 방식을 fs모듈로 가져오는 방식으로 변경
